+++ src/nrepl/jack-in.ts
@@ -204,6 +204,10 @@ async function getJackInTerminalOptions(
 
   const projectType = projectTypes.getProjectTypeForName(projectTypeName);
 
+  if (!projectType?.commandLine) {
+    throw new Error(`Project type ${projectTypeName} does not support Jack-in.`);
+  }
+
   const commandLineInfo = await projectType.commandLine(projectConnectSequence, selectedCljsType);
 
   let args: string[] = commandLineInfo.args;